# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'HaloSitek Core API'
  description: 'OpenAPI metadata for HaloSitek Backoffice API.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api/v1'
    description: 'Local Development Server'
paths:
...
  '/chat/conversations/{conversationId}/messages':
    get:
      tags:
        - Chat
      summary: 'List messages'
      description: 'Mengambil daftar pesan pada conversation tertentu.'
      operationId: 318164aae8b37e8a0c5d03c99a230c5b
      parameters:
        -
          name: conversationId
          in: path
          required: true
          schema:
            type: string
        -
          name: per_page
          in: query
          required: false
          schema:
            type: integer
            example: 20
            maximum: 100
            minimum: 1
      responses:
        200:
          description: 'Daftar pesan berhasil diambil'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'Daftar pesan berhasil diambil.'
                data:
                  -
                    id: 01J2MESSAGE001
                    conversation_id: 01J2CHATCONVERSATION001
                    user_id: 01J2USERA
                    body: Halo!
                    attachment: null
                    read_at: null
                    is_mine: true
                    sender:
                      id: 01J2USERA
                      name: Budi
                      email: budi@example.com
                    created_at: '2026-04-19T10:00:00+00:00'
                    updated_at: '2026-04-19T10:00:00+00:00'
                meta:
                  current_page: 1
                  last_page: 1
                  per_page: 20
                  total: 1
                links:
                  first_page_url: 'http://localhost:8000/api/v1/chat/conversations/01J2CHATCONVERSATION001/messages?page=1'
                  last_page_url: 'http://localhost:8000/api/v1/chat/conversations/01J2CHATCONVERSATION001/messages?page=1'
                  next_page_url: null
                  prev_page_url: null
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
  /chat/messages:
    post:
      tags:
        - Chat
      summary: 'Send message'
      description: 'Mengirim pesan teks ke conversation. Untuk role user, sistem memproses AI reply (sync untuk teks, async/queue untuk request gambar).'
      operationId: 5544065cc5106aa1f9a34cfc8762cfe4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - conversation_id
                - body
              properties:
                conversation_id:
                  type: string
                  example: 01J2CHATCONVERSATION001
                body:
                  type: string
                  example: 'Halo, kabar kamu gimana?'
                attachment:
                  type: string
                  example: null
                  nullable: true
              type: object
      responses:
        201:
          description: 'Pesan berhasil diproses dan AI reply sinkron dikembalikan'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 201
                message: 'Pesan berhasil dikirim.'
                data:
                  id: 01J2MESSAGE001
                  conversation_id: 01J2CHATCONVERSATION001
                  user_id: 01J2USERA
                  role: assistant
                  type: text
                  content: 'Ini jawaban AI konsultasi.'
                  body: 'Ini jawaban AI konsultasi.'
                  attachment: null
                  read_at: null
                  is_mine: false
                  sender:
                    id: 01J2USERA
                    name: Budi
                    email: budi@example.com
                  created_at: '2026-04-19T10:00:00+00:00'
                  updated_at: '2026-04-19T10:00:00+00:00'
        202:
          description: 'Permintaan AI sedang diproses melalui queue'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 202
                message: 'Permintaan AI sedang diproses.'
                data:
                  status: processing
                  conversation_id: 01J2CHATCONVERSATION001
                  user_message_id: 01J2MESSAGE001
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        422:
          $ref: '#/components/responses/ValidationError'
        503:
          $ref: '#/components/responses/ServiceUnavailableError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
  /chat/ai/messages:
    post:
      tags:
        - Chat
      summary: 'Send AI message (legacy endpoint)'
      description: 'Mengirim pesan ke AI service dengan history berbasis user (endpoint legacy).'
      operationId: 709e48a6621b39699d0ac46d458d9504
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - message
              properties:
                message:
                  type: string
                  example: 'Buatkan konsep rumah tropis.'
                  maxLength: 2000
              type: object
      responses:
        200:
          description: 'AI response berhasil'
          content:
            application/json:
              schema: []
              example:
                type: text
                content: 'Ini adalah jawaban AI.'
                prompt_used: null
        401:
          $ref: '#/components/responses/UnauthorizedError'
        422:
          $ref: '#/components/responses/ValidationError'
        503:
          $ref: '#/components/responses/ServiceUnavailableError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
  '/chat/conversations/{conversationId}/read':
    post:
      tags:
        - Chat
      summary: 'Mark conversation as read'
      description: 'Menandai seluruh pesan lawan bicara pada conversation sebagai sudah dibaca dan update last_read_at user.'
      operationId: 52396d03c385f258bf853fe18579ce15
      parameters:
        -
          name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Pesan berhasil ditandai sudah dibaca'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'Pesan berhasil ditandai sudah dibaca.'
                data:
                  id: 01J2CHATCONVERSATION001
                  name: null
                  is_group: false
                  participant_ids:
                    - 01J2USERA
                    - 01J2USERB
                  last_read_at: '2026-04-19T10:05:00+00:00'
                  created_at: '2026-04-19T10:00:00+00:00'
                  updated_at: '2026-04-19T10:05:00+00:00'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
...
components:
...
  responses:
    UnauthorizedError:
      description: 'Unauthorized - missing or invalid access token.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 'Unauthorized. Access token is missing or invalid.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
    ForbiddenError:
      description: 'Forbidden - user does not have permission for this action.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: 'Forbidden. You are not allowed to perform this action.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
    NotFoundError:
      description: 'Not Found - requested resource was not found.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: 'Resource not found.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
...
    ValidationError:
      description: 'Validation Error - request input is invalid.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 422
                  message:
                    type: string
                    example: 'Validation error.'
                  errors:
                    type: object
                    example:
                      email:
                        - 'The email field is required.'
                      password:
                        - 'The password must be at least 8 characters.'
                type: object
    ServerError:
      description: 'Internal Server Error - an unexpected server error occurred.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: 'Internal server error.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
    ServiceUnavailableError:
      description: 'Service Unavailable - dependent service is temporarily unavailable.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 503
                  message:
                    type: string
                    example: 'Service unavailable.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
  securitySchemes:
    BearerAuth:
      type: http
      description: 'Use Sanctum access token in Authorization header as: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...